Split macOS app source files
