+++ macos/AgentCLI/Sources/AgentCLI/CommandResult.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct CommandResult {
+    let exitCode: Int32
+    let output: String
+}
